libfreefare: Remove an irrelevant FIXME.
The check is already done by the PICC and we may not have access to the
required information to do the check from the library. Additionally, the
mifare_desfire_change_key() would require different keys to operate if we
detected an error, so no recovery of any kind is possible.



git-svn-id: http://nfc-tools.googlecode.com/svn/trunk@540 ba5c3050-9c55-11de-a262-85a698460a8e